Hands on Activity #2/google_sample_function/.ipynb_checkpoints/requirements-checkpoint.txt
pandas
datetime
google-api-python-client==1.7.11
oauth2client
pandas_gbq